KloudwitKloudManager/requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0

# Cloud Provider SDKs
boto3==1.34.20  # AWS
azure-identity==1.15.0
azure-mgmt-compute==30.5.0
azure-mgmt-resource==23.0.1
azure-mgmt-storage==21.1.0
azure-mgmt-network==25.2.0
azure-mgmt-web==10.1.0
azure-mgmt-sql==3.0.1
azure-mgmt-policyinsights==1.1.0b4
azure-mgmt-costmanagement==4.0.1
google-cloud-compute==1.15.0
google-cloud-storage==2.14.0
google-auth==2.26.2

# Database
SQLAlchemy==2.0.25

# Utilities
python-dotenv==1.0.0
requests==2.31.0
cryptography==42.0.0

# Scheduled Tasks
APScheduler==3.10.4

# Data Processing
pandas==2.1.4
